+++ bundles/ch.elexis.core.jpa.entities/src/ch/elexis/core/jpa/entities/KontaktAdressJoint.java
@@ -1,7 +1,5 @@
 package ch.elexis.core.jpa.entities;
 
-import java.math.BigInteger;
-
 import javax.persistence.Column;
 import javax.persistence.Convert;
 import javax.persistence.Entity;
@@ -29,7 +27,7 @@ public class KontaktAdressJoint implements EntityWithId, EntityWithDeleted {
 	public static final String QUERY_findAllIDisOtherKontakt = "QUERY_findAllIDisOtherKontakt";
 
 	// Transparently updated by the EntityListener
-	protected BigInteger lastupdate;
+	protected Long lastupdate;
 	
 	@Id
 	@GeneratedValue(generator = "system-uuid")
@@ -116,12 +114,12 @@ public class KontaktAdressJoint implements EntityWithId, EntityWithDeleted {
 	}
 	
 	@Override
-	public BigInteger getLastupdate(){
+	public Long getLastupdate(){
 		return lastupdate;
 	}
 	
 	@Override
-	public void setLastupdate(BigInteger lastupdate){
+	public void setLastupdate(Long lastupdate){
 		this.lastupdate = lastupdate;
 	}
 }
