+++ bundles/ch.elexis.core.jpa.entities/src/ch/elexis/core/jpa/entities/Labor2009Tarif.java
@@ -1,6 +1,5 @@
 package ch.elexis.core.jpa.entities;
 
-import java.math.BigInteger;
 import java.time.LocalDate;
 
 import javax.persistence.Column;
@@ -25,7 +24,7 @@ public class Labor2009Tarif implements EntityWithId, EntityWithDeleted {
 	public static final String CODESYSTEM_NAME = "EAL 2009";
 
 	// Transparently updated by the EntityListener
-	protected BigInteger lastupdate;
+	protected Long lastupdate;
 	
 	@Id
 	@GeneratedValue(generator = "system-uuid")
@@ -168,12 +167,12 @@ public class Labor2009Tarif implements EntityWithId, EntityWithDeleted {
 	}
 	
 	@Override
-	public BigInteger getLastupdate(){
+	public Long getLastupdate(){
 		return lastupdate;
 	}
 	
 	@Override
-	public void setLastupdate(BigInteger lastupdate){
+	public void setLastupdate(Long lastupdate){
 		this.lastupdate = lastupdate;
 	}
 }
