+++ bundles/ch.elexis.core.jpa.entities/src/ch/elexis/core/jpa/entities/StockEntry.java
@@ -1,8 +1,6 @@
 package ch.elexis.core.jpa.entities;
 
 
-import java.math.BigInteger;
-
 import javax.persistence.Column;
 import javax.persistence.Convert;
 import javax.persistence.Entity;
@@ -35,7 +33,7 @@ import ch.elexis.core.jpa.entities.listener.EntityWithIdListener;
 public class StockEntry implements EntityWithId, EntityWithDeleted {
 	
 	// Transparently updated by the EntityListener
-	protected BigInteger lastupdate;
+	protected Long lastupdate;
 	
 	@Id
 	@GeneratedValue(generator = "system-uuid")
@@ -202,12 +200,12 @@ public class StockEntry implements EntityWithId, EntityWithDeleted {
 	}
 	
 	@Override
-	public BigInteger getLastupdate(){
+	public Long getLastupdate(){
 		return lastupdate;
 	}
 	
 	@Override
-	public void setLastupdate(BigInteger lastupdate){
+	public void setLastupdate(Long lastupdate){
 		this.lastupdate = lastupdate;
 	}
 }
