+++ bundles/ch.elexis.core.jpa.entities/src/ch/elexis/core/jpa/entities/TarmedKumulation.java
@@ -1,6 +1,5 @@
 package ch.elexis.core.jpa.entities;
 
-import java.math.BigInteger;
 import java.time.LocalDate;
 
 import javax.persistence.Column;
@@ -27,7 +26,7 @@ public class TarmedKumulation implements EntityWithId, EntityWithDeleted {
 	public static final String TYP_EXCLUSIVE = "X";
 
 	// Transparently updated by the EntityListener
-	protected BigInteger lastupdate;
+	protected Long lastupdate;
 	
 	@Id
 	@GeneratedValue(generator = "system-uuid")
@@ -169,12 +168,12 @@ public class TarmedKumulation implements EntityWithId, EntityWithDeleted {
 	}
 	
 	@Override
-	public BigInteger getLastupdate(){
+	public Long getLastupdate(){
 		return lastupdate;
 	}
 	
 	@Override
-	public void setLastupdate(BigInteger lastupdate){
+	public void setLastupdate(Long lastupdate){
 		this.lastupdate = lastupdate;
 	}
